+++ lib/radiator_web/templates/admin/podcast/index.html.eex
@@ -1,34 +1,34 @@
-<div class="container mx-auto mb-6 px-8">
+<div class="container mb-6 px-8">
     <div class="mb-6">
-      <h1 class="text-3xl font-bold mb-1 text-blue-900">Hi, Happy Podcasting!</h1>
-      <span class="text-gray-600 font-normal text-lg">Let's spark some joy and enlightenment, shall we?</span>
+      <h1 class="text-3xl font-bold mb-1 text-blue-900">Podcasts in <%= @current_network.title %></h1>
     </div>
 
     <div class="flex justify-center md:justify-start flex-wrap">
       <%= for podcast <- @podcasts do %>
-      <a href="<%= Routes.admin_network_podcast_path(@conn, :show, podcast.network_id, podcast) %>" class="no-underline text-black w-48 bg-white rounded flex flex-col shadow mr-8 mb-8">
+        <%= link to: Routes.admin_network_podcast_path(@conn, :show, podcast.network_id, podcast),
+        class: "no-underline text-black w-48 bg-white rounded flex flex-col shadow mr-8 mb-8" do %>
         
-        <div
-          class="w-48 h-48 bg-cover shadow-inner"
-          style="background-image: url(<%= podcast_image_url(podcast) %>)"
-        ></div>
+          <div
+            class="w-48 h-48 bg-cover shadow-inner"
+            style="background-image: url(<%= podcast_image_url(podcast) %>)"
+          ></div>
 
-        <div class="p-3 h-32 overflow-hidden">
-          <h1 class="text-base mb-1"><%= podcast.title %></h1>
-          <p class="text-xs text-gray-600"><%= shorten_string(podcast.subtitle, 140) %></p>
-        </div>
+          <div class="p-3 h-32 overflow-hidden">
+            <h1 class="text-base mb-1"><%= podcast.title %></h1>
+            <p class="text-xs text-gray-600"><%= shorten_string(podcast.subtitle, 140) %></p>
+          </div>
 
-        <div class="flex items-center p-3 border-t border-gray-500 border-solid text-xs text-gray-600">
-          <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" class="w-4 mr-3 opacity-75 icon-collection"><rect width="20" height="12" x="2" y="10" class="primary" rx="2"></rect><path class="secondary" d="M20 8H4c0-1.1.9-2 2-2h12a2 2 0 0 1 2 2zm-2-4H6c0-1.1.9-2 2-2h8a2 2 0 0 1 2 2z"></path></svg>
-          <span><%= ngettext "One Episode", "%{count} Episodes", podcast.episode_count %></span>
-        </div>
+          <div class="flex items-center p-3 border-t border-gray-500 border-solid text-xs text-gray-600">
+            <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" class="w-4 mr-3 opacity-75 icon-collection"><rect width="20" height="12" x="2" y="10" class="primary" rx="2"></rect><path class="secondary" d="M20 8H4c0-1.1.9-2 2-2h12a2 2 0 0 1 2 2zm-2-4H6c0-1.1.9-2 2-2h8a2 2 0 0 1 2 2z"></path></svg>
+            <span><%= ngettext "One Episode", "%{count} Episodes", podcast.episode_count %></span>
+          </div>
 
-        <div class="flex items-center p-3 pt-0 border-gray-500 border-solid text-xs text-gray-600">
-          <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" class="w-4 mr-3 opacity-75 icon-trending-up"><path class="primary" d="M3.7 20.7a1 1 0 1 1-1.4-1.4l6-6a1 1 0 0 1 1.4 0l3.3 3.29 4.3-4.3a1 1 0 0 1 1.4 1.42l-5 5a1 1 0 0 1-1.4 0L9 15.4l-5.3 5.3z"></path><path class="secondary" d="M16.59 8l-2.3-2.3A1 1 0 0 1 15 4h6a1 1 0 0 1 1 1v6a1 1 0 0 1-1.7.7L18 9.42l-4.3 4.3a1 1 0 0 1-1.4 0L9 10.4l-5.3 5.3a1 1 0 1 1-1.4-1.42l6-6a1 1 0 0 1 1.4 0l3.3 3.3L16.59 8z"></path></svg>
-          <span>123.456 Downloads</span>
-        </div>
+          <%# <div class="flex items-center p-3 pt-0 border-gray-500 border-solid text-xs text-gray-600">
+            <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" class="w-4 mr-3 opacity-75 icon-trending-up"><path class="primary" d="M3.7 20.7a1 1 0 1 1-1.4-1.4l6-6a1 1 0 0 1 1.4 0l3.3 3.29 4.3-4.3a1 1 0 0 1 1.4 1.42l-5 5a1 1 0 0 1-1.4 0L9 15.4l-5.3 5.3z"></path><path class="secondary" d="M16.59 8l-2.3-2.3A1 1 0 0 1 15 4h6a1 1 0 0 1 1 1v6a1 1 0 0 1-1.7.7L18 9.42l-4.3 4.3a1 1 0 0 1-1.4 0L9 10.4l-5.3 5.3a1 1 0 1 1-1.4-1.42l6-6a1 1 0 0 1 1.4 0l3.3 3.3L16.59 8z"></path></svg>
+            <span>123.456 Downloads</span>
+          </div> %>
 
-      </a>        
+        <% end %>
       <% end %>
 
       <%= if has_manage_permission_for_network(@authenticated_user, @current_network) do %>
