+++ lib/radiator_web/templates/admin/podcast/show.html.eex
@@ -2,48 +2,51 @@
 
   <div class="mb-6 flex w-full justify-between">
     
-    <div class="flex justify-between w-full">
+    <div class="justify-between w-full">
     
-      <div class="w-3/4">
-        <div class="flex">
-          <h1 class="text-3xl font-bold mb-1 text-blue-900 mr-4">
-            <%= @podcast.title %>
-            <%= if @podcast.published_at && @podcast.slug do %>
-            <span class="text-gray-600 font-normal text-lg">
-              <%= link to: Routes.feed_path(@conn, :show, @podcast.slug), class: "txt-lnk" do %>
-                RSS Feed
-              <% end %>   
-            </span>
-            <span class="text-gray-600 font-normal text-lg">
-            |
-            </span>
-            <span class="text-gray-600 font-normal text-lg">
-              <%= link to: Routes.episode_path(@conn, :index, @podcast.slug), class: "no-underline text-blue-500", target: "_blank" do %>
-                Public Page
-              <% end %>   
-            </span>
-            <% end %>
-          </h1>
-          <%= link to: Routes.admin_network_podcast_path(@conn, :edit, @podcast.network_id, @podcast), class: "no-underline text-blue-500" do %>
-          <div class="pill">edit</div>
-          <% end %>      
-        </div>
-        <span class="text-gray-600 font-normal text-lg">
-          <%= @podcast.subtitle %>
-        </span>
-      </div>
-
-      <div class="flex w-full h-12 justify-end mb-2">
-        <%= link to: Routes.admin_network_podcast_episode_path(@conn, :new, @podcast.network_id, @podcast), class: "btn inline-flex items-center" do %>
-          <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" class="w-6 h-6 mr-2 icon-add"><path fill="white" d="M17 11a1 1 0 0 1 0 2h-4v4a1 1 0 0 1-2 0v-4H7a1 1 0 0 1 0-2h4V7a1 1 0 0 1 2 0v4h4z"></path></svg>
+      <div class="">
+        <div class="inline h-12 justify-end mb-2 ml-2 float-right">
+      <%= link to: Routes.admin_network_podcast_path(@conn, :edit, @podcast.network_id, @podcast), class: "btn btn-wicon mr-2" do %>
+        <svg xmlns="http://www.w3.org/2000/svg" viewBox="-4 -4 30 30" class="w-6 h-6 fill-current mr-2"><path d="M4 14a1 1 0 0 1 .3-.7l11-11a1 1 0 0 1 1.4 0l3 3a1 1 0 0 1 0 1.4l-11 11a1 1 0 0 1-.7.3H5a1 1 0 0 1-1-1v-3z"/></svg>
+        <span>Edit</span>
+      <% end %> 
+        <%= link to: Routes.admin_network_podcast_episode_path(@conn, :new, @podcast.network_id, @podcast), class: "btn btn-wicon" do %>
+          <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" class="w-6 h-6 mr-2 fill-current"><path d="M17 11a1 1 0 0 1 0 2h-4v4a1 1 0 0 1-2 0v-4H7a1 1 0 0 1 0-2h4V7a1 1 0 0 1 2 0v4h4z"></path></svg>
           <span>Create Episode</span>
         <% end %>
+        </div>
+
+
+        <h1 class="text-3xl font-bold mb-1 mt-2 text-blue-900 mr-4">
+          <%= @podcast.title %>
+        </h1>
+        <p class="text-gray-600 font-normal text-lg mb-2">
+          <%= @podcast.subtitle %>
+        </p>
+        <%= if @podcast.published_at && @podcast.slug do %>
+          <p>
+          <span class="text-gray-600 font-normal text-lg">
+            <%= link "RSS Feed", to: Routes.feed_path(@conn, :show, @podcast.slug), class: "txt-lnk"%>   
+          </span>
+          <span class="text-gray-600 font-normal text-lg">
+          &nbsp;|&nbsp;
+          </span>
+          <span class="text-gray-600 font-normal text-lg">
+            <%= link "Public Page", to: Routes.episode_path(@conn, :index, @podcast.slug), class: "txt-lnk", target: "_blank" %>   
+          </span>
+          </p>
+        <% end %>
+
       </div>
+
+      
 
     </div>
 </div>
 
-  <h2 class="text-gray-700 mb-2">Drafts</h2>
+  <hr class="border-b border-secondary" />
+
+  <h2 class="text-gray-700 text-xl font-bold mb-2">Drafts</h2>
 
   <%= for episode <- @draft_episodes do %>
     <%= link to: Routes.admin_network_podcast_episode_path(@conn, :show, @podcast.network_id, @podcast, episode), class: "block no-underline card mb-1 hover:bg-highlight" do %>
@@ -64,7 +67,9 @@
     <% end %>
   <% end %>
 
-  <h2 class="text-gray-700 mb-2">Published</h2>
+  <hr class="border-b border-secondary" />
+
+  <h2 class="text-gray-700 text-xl font-bold mb-2">Published</h2>
 
   <%= for episode <- @published_episodes do %>
     <%= link to: Routes.admin_network_podcast_episode_path(@conn, :show, @podcast.network_id, @podcast, episode), class: "block no-underline card mb-1 hover:bg-highlight" do %>
