+++ lib/radiator_web/templates/page/index.html.eex
@@ -1,10 +1,10 @@
 <div class="container mx-auto mt-8">
 
-  <div class="max-w-lg rounded overflow-hidden shadow-lg mb-8 bg-white">
+  <div class="max-w-2xl mx-auto rounded overflow-hidden shadow-lg mb-8 bg-white">
     <div class="px-6 py-6">
       <header class=" mb-8">
         <h1 class="font-bold text-gray-800 text-2xl mb-1">Welcome to Podlove Radiator</h1>
-        <h2 class="text-xl text-secondary">Podcast hosting for the next century of the internet.</h2>
+        <h2 class="text-xl text-secondary">Podcast hosting for the next century.</h2>
       </header>
 
       <div class="mb-6">
@@ -22,7 +22,7 @@
 
         <ul>
           <li>
-            <a href="https://github.com/podlove/radiator" class="text-primary hover:underline"><tt>podlove/radiator</tt> on GitHub</a>
+            <a href="https://github.com/podlove/radiator" class="text-primary hover:underline"> <tt>podlove/radiator</tt> on GitHub</a>
           </li>
         </ul>
       </div>
