+++ lib/radiator_web/templates/page/nav.html.eex
@@ -2,9 +2,9 @@
   <div class="w-full bg-secondary h-1"></div>
 </div>
 <div class="flex mb-10 text-white">
-  <div class="w-full bg-primary h-16 flex items-center px-5">
-    <span class="text-2xl font-thin">
-      Radiator 🔥
+  <div class="w-full bg-primary h-16 flex items-center px-4">
+    <span class="text-2xl flex-shrink-0 mr-2">
+      Radiator&nbsp;<svg class="w-10 h-10 inline" viewBox="0 0 900 980" xmlns="http://www.w3.org/2000/svg" fill-rule="evenodd" clip-rule="evenodd" stroke-linejoin="round" stroke-miterlimit="1.414"><path d="M944.137 231.69a70.12 70.12 0 0 1 .744 10.094v588.865c0 8.5-.486 16.426-1.408 23.8l-36.79-35.501V266.727l37.454-35.037zm-74.582 551.417l-38.197-36.857V337.194l38.197-35.726v481.639zm-75.464-72.829l-38.188-36.857V407.79l38.188-35.726v338.214zM48.643 231.69l36.78 35.502v552.22L47.98 854.45a70.165 70.165 0 0 1-.743-10.094V255.491c0-8.5.486-16.427 1.407-23.8zm112.107 108.2v409.056l-38.188 35.726v-481.64l38.188 36.857zm75.464 72.829v265.63l-38.188 35.727V375.861l38.188 36.857z" class="primary" fill-rule="nonzero"/><path d="M617.43 47.244H372.497c-50.377 0-91.227 40.85-91.227 91.236v509.526c0 50.386 40.85 91.232 91.227 91.232H617.43c50.39 0 91.232-40.846 91.232-91.232V138.48c0-50.386-40.842-91.236-91.232-91.236" fill="#fff" fill-rule="nonzero"/><path d="M617.43 47.244H372.497c-50.377 0-91.227 40.85-91.227 91.236v509.526c0 50.386 40.85 91.232 91.227 91.232H617.43c50.39 0 91.232-40.846 91.232-91.232V138.48c0-50.386-40.842-91.236-91.232-91.236m0 38.729c28.952 0 52.503 23.555 52.503 52.507v509.526c0 28.952-23.551 52.503-52.503 52.503H372.497c-28.947 0-52.498-23.55-52.498-52.503V138.48c0-28.952 23.551-52.507 52.498-52.507H617.43" fill="#181716" fill-rule="nonzero"/><path d="M494.604 397.575c-76.088 0-137.99 61.906-137.99 137.995 0 76.093 61.902 137.998 137.99 137.998 76.093 0 137.999-61.905 137.999-137.998 0-76.09-61.906-137.995-137.999-137.995m0 38.729c54.831 0 99.27 44.439 99.27 99.266 0 54.822-44.439 99.27-99.27 99.27-54.813 0-99.261-44.448-99.261-99.27 0-54.827 44.448-99.266 99.261-99.266m55.443-319.035c44.938.852 84.893 42.18 82.595 88.008-1.012 20.195-9.63 39.744-23.892 54.457l-115.226 115.29-115.217-115.299c-4.618-4.763-5.81-6.39-8.314-9.88-31.611-44.039-7.436-118.118 50.2-130.666 25.424-5.534 53.14 1.536 73.004 18.766 0 0 29.098-20.354 55.055-20.68.9-.005.895-.005 1.795.004zm-112.554 38.724c-34.22.65-59.668 47.602-31.586 76.562l87.622 87.681 87.617-87.668c25.668-26.482 3.705-84.437-40.863-75.745-12.325 2.402-23.358 10.225-29.907 20.983l-16.843 28.173c-14.076-23.539-28.238-49.65-55.064-49.986-.486-.004-.486 0-.976 0z" fill="#181716" fill-rule="nonzero"/><ellipse cx="73.33" cy="47.329" rx="8.258" ry="8.258" fill="#181716" transform="matrix(-4.30318 0 0 4.30318 810.165 331.905)"/><path fill="none" d="M-.002 0h992.125v897.64H-.002z"/></svg>
     </span>
     <ul class="flex flex-row items-center h-full text-lg font-thin ml-4">
       <%= if Guardian.Plug.current_resource(@conn) do %>
@@ -17,37 +17,42 @@
             <%= link  @current_network.title , to: Routes.admin_network_podcast_path(@conn, :index, @current_network) %>
           </li>
           <%= if assigns[:podcast] do %>
-          <%= render("nav_chevron.html") %>
-          <li class="nav-link">
-            <%= link @podcast.title, to: Routes.admin_network_podcast_path(@conn, :show, @podcast.network_id, @podcast) %>
-          </li>      
+            <%= render("nav_chevron.html") %>
+            <li class="nav-link">
+              <%= link @podcast.title, to: Routes.admin_network_podcast_path(@conn, :show, @podcast.network_id, @podcast) %>
+            </li>      
           <% end %>
           <%= if assigns[:episode] do %>
-          <%= render("nav_chevron.html") %>
-          <li class="nav-link">
-            <%= link @episode.podcast.title, to: Routes.admin_network_podcast_path(@conn, :show, @episode.podcast.network_id, @episode.podcast) %>
-          </li>
-          <%= render("nav_chevron.html") %>
-          <li class="nav-link">
-            <%= link @episode.title, to: Routes.admin_network_podcast_episode_path(@conn, :show, @episode.podcast.network_id, @episode.podcast, @episode) %>
-          </li>
+            <%= render("nav_chevron.html") %>
+            <li class="nav-link truncate-fade">
+              <%= link shorten_string(@episode.podcast.title, 32), to: Routes.admin_network_podcast_path(@conn, :show, @episode.podcast.network_id, @episode.podcast) %>
+            </li>
+            <%= render("nav_chevron.html") %>
+            <li class="nav-link">
+              <%= link @episode.title, to: Routes.admin_network_podcast_episode_path(@conn, :show, @episode.podcast.network_id, @episode.podcast, @episode) %>
+            </li>
           <% end %>        
         <% end %>
       <% end %> 
     </ul>
 
-    <span class="font-thin ml-auto nav-link" style="margin-left:auto">
+    <div class="inline-flex ml-auto items-center ml-3" style="margin-left:auto">
       <%= case Guardian.Plug.current_resource(@conn) do %>
-        <% %Radiator.Auth.User{ name: name } -> %>
-          <li class="nav-link">
-            <%= link "My Profile", to: Routes.admin_user_settings_path(@conn, :index) %>
+        <% %Radiator.Auth.User{ name: name, display_name: display_name } -> %>
+          <li class="nav-link text-center mx-2">
+            <%= link to: Routes.admin_user_settings_path(@conn, :index) do %>
+              <%= content_tag :span, name %>
+              <%= if name != display_name do %>
+                <br /><%= content_tag :span, display_name, class: "text-sm font-thin"  %>
+              <% end %>
+            <% end %>
           </li>
-          <li class="nav-link">
-            <%= link "#{name} (logout)", to: Routes.login_path(@conn, :logout) %>
+          <li class="nav-link pl-2">
+            <%= link "Logout", to: Routes.login_path(@conn, :logout) %>
           </li>
-        <% _ -> %> <%= link "login", to: Routes.login_path(@conn, :login_form) %>
+        <% _ -> %> <%= link "Login", to: Routes.login_path(@conn, :login_form) %>
       <% end %>      
-    </span>
+    </div>
 
   </div>
 </div>
