+++ lib/radiator_web/templates/page/nav_chevron.html.eex
@@ -1,4 +1,4 @@
-<li class="nav-chevron">
+<li class="nav-chevron flex-shrink-0">
     <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" class="w-6 icon-chevron-right-circle">
         <path class="primary" d="M10.3 8.7a1 1 0 0 1 1.4-1.4l4 4a1 1 0 0 1 0 1.4l-4 4a1 1 0 0 1-1.4-1.4l3.29-3.3-3.3-3.3z"></path>
     </svg>
